+++ VetCrm/Controllers/RacaController.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -19,42 +15,26 @@ namespace VetCrm.Controllers
             _context = context;
         }
 
-        // GET: Raca
         public async Task<IActionResult> Index()
         {
             var vetCrmContext = _context.Racas.Include(r => r.Especie);
             return View(await vetCrmContext.ToListAsync());
         }
 
-        // GET: Raca/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var raca = await _context.Racas
-                .Include(r => r.Especie)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (raca == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var raca = await _context.Racas.Include(r => r.Especie).FirstOrDefaultAsync(m => m.Id == id);
+            if (raca == null) return NotFound();
             return View(raca);
         }
 
-        // GET: Raca/Create
         public IActionResult Create()
         {
-            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Id");
+            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Nome");
             return View();
         }
 
-        // POST: Raca/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Nome,EspecieId")] Raca raca)
@@ -65,38 +45,24 @@ namespace VetCrm.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Id", raca.EspecieId);
+            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Nome", raca.EspecieId);
             return View(raca);
         }
 
-        // GET: Raca/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var raca = await _context.Racas.FindAsync(id);
-            if (raca == null)
-            {
-                return NotFound();
-            }
-            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Id", raca.EspecieId);
+            if (raca == null) return NotFound();
+            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Nome", raca.EspecieId);
             return View(raca);
         }
 
-        // POST: Raca/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,EspecieId")] Raca raca)
         {
-            if (id != raca.Id)
-            {
-                return NotFound();
-            }
+            if (id != raca.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -107,51 +73,29 @@ namespace VetCrm.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!RacaExists(raca.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!RacaExists(raca.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Id", raca.EspecieId);
+            ViewData["EspecieId"] = new SelectList(_context.Especies, "Id", "Nome", raca.EspecieId);
             return View(raca);
         }
 
-        // GET: Raca/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var raca = await _context.Racas
-                .Include(r => r.Especie)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (raca == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var raca = await _context.Racas.Include(r => r.Especie).FirstOrDefaultAsync(m => m.Id == id);
+            if (raca == null) return NotFound();
             return View(raca);
         }
 
-        // POST: Raca/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var raca = await _context.Racas.FindAsync(id);
-            if (raca != null)
-            {
-                _context.Racas.Remove(raca);
-            }
-
+            if (raca != null) _context.Racas.Remove(raca);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
